+++ app/(tabs)/_layout.tsx
@@ -7,7 +7,7 @@ import { IconSymbol } from '@/components/ui/IconSymbol';
 import TabBarBackground from '@/components/ui/TabBarBackground';
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
-import { ProfileScreen } from '@/screens/ProfileScreen';
+import { ProfileScreen } from './profile';
 
 export default function TabLayout() {
   const colorScheme = useColorScheme();
