+++ app/(tabs)/explore.tsx
@@ -6,7 +6,7 @@ import ParallaxScrollView from '@/components/ParallaxScrollView';
 import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
 import { IconSymbol } from '@/components/ui/IconSymbol';
-import { ProfileScreen } from '@/screens/ProfileScreen';
+
 
 export default function TabTwoScreen() {
   return (
